chore: move workflow/webhook logs test into their own file
