+++ src/services/webhook.js
@@ -313,7 +313,7 @@ function start ({ communication }) {
           query: 'inList'
         },
         statuses: {
-          dbField: 'statusCode',
+          dbField: 'status',
           value: status,
           transformValue: 'array',
           query: 'inList'
