+++ test/integration/api/webhook.spec.js
@@ -4,7 +4,6 @@ const test = require('ava')
 const request = require('supertest')
 const express = require('express')
 const bodyParser = require('body-parser')
-const _ = require('lodash')
 
 const userServer = express()
 let userServerPort
@@ -15,7 +14,6 @@ const { before, beforeEach, after } = require('../../lifecycle')
 const { getAccessTokenHeaders } = require('../../auth')
 const {
   testEventDelay,
-  computeDate,
 
   checkCursorPaginationScenario,
   checkCursorPaginatedListObject,
@@ -24,85 +22,9 @@ const { apiVersions } = require('../../../src/versions')
 
 let userWebhookUrl
 
-const defaultTestDelay = 4000
-
-let createdWebhooks
+const getIds = (elements) => elements.map(e => e.id)
 
 /* eslint-disable no-template-curly-in-string */
-
-async function createWebhookLogs (t) {
-  if (createdWebhooks) return createdWebhooks
-
-  const authorizationHeaders = await getAccessTokenHeaders({
-    t,
-    permissions: [
-      'webhookLog:list:all',
-      'webhook:create:all',
-      'category:create:all',
-      'entry:create:all',
-      'message:create:all',
-    ]
-  })
-
-  // should create webhooks that listen to events that are not triggered by any tests below
-
-  const { body: messageWebhook } = await request(t.context.serverUrl)
-    .post('/webhooks')
-    .set(authorizationHeaders)
-    .send({
-      name: 'Webhook for message creation',
-      event: 'message__created',
-      targetUrl: userWebhookUrl + 'messageCreation',
-    })
-    .expect(200)
-
-  const { body: entryWebhook } = await request(t.context.serverUrl)
-    .post('/webhooks')
-    .set(authorizationHeaders)
-    .send({
-      name: 'Webhook for entry creation',
-      event: 'entry__created',
-      targetUrl: userWebhookUrl + 'entryCreation',
-    })
-    .expect(200)
-
-  createdWebhooks = _.keyBy([
-    messageWebhook,
-    entryWebhook,
-  ], 'event')
-
-  await request(t.context.serverUrl)
-    .post('/messages')
-    .set(authorizationHeaders)
-    .send({
-      topicId: 'ast_2l7fQps1I3a1gJYz2I3a',
-      receiverId: 'user-external-id',
-      content: 'Good',
-    })
-    .expect(200)
-
-  await request(t.context.serverUrl)
-    .post('/entries')
-    .set(authorizationHeaders)
-    .send({
-      collection: 'someCollection',
-      locale: 'en-US',
-      name: 'nameExample',
-      fields: {
-        title: 'Random title',
-        content: 'Random content',
-        nestedContent: {
-          random1: {
-            random2: 'hello'
-          },
-          random3: 'bye'
-        }
-      }
-    })
-    .expect(200)
-
-  await new Promise(resolve => setTimeout(resolve, defaultTestDelay))
-}
 
 test.before(async (t) => {
   await before({ name: 'webhook' })(t)
@@ -133,8 +55,6 @@ test.before(async (t) => {
       resolve()
     })
   })
-
-  await createWebhookLogs(t)
 })
 // test.beforeEach(beforeEach()) // concurrent tests are much faster
 test.after(async (t) => {
@@ -213,6 +133,7 @@ test('creates a webhook', async (t) => {
     permissions: [
       'webhook:create:all',
       'webhook:read:all',
+      'webhookLog:list:all',
       'asset:create:all',
       'platformData:edit:all'
     ]
@@ -257,6 +178,13 @@ test('creates a webhook', async (t) => {
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: webhookLogsAfterCall } } = await request(t.context.serverUrl)
+    .get(`/webhook-logs?webhookId=${webhook.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(webhookAfterCall.logs), getIds(webhookLogsAfterCall))
+
   t.is(webhookAfterCall.logs.length, 1)
   t.is(webhookAfterCall.logs[0].status, 'success')
   t.is(webhookAfterCall.logs[0].metadata.eventObjectId, asset.id)
@@ -268,6 +196,7 @@ test('creates a webhook with specified API version', async (t) => {
     permissions: [
       'webhook:create:all',
       'webhook:read:all',
+      'webhookLog:list:all',
       'assetType:create:all',
       'platformData:edit:all'
     ]
@@ -315,6 +244,13 @@ test('creates a webhook with specified API version', async (t) => {
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: webhookLogsAfterCall } } = await request(t.context.serverUrl)
+    .get(`/webhook-logs?webhookId=${webhook.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(webhookAfterCall.logs), getIds(webhookLogsAfterCall))
+
   t.is(webhookAfterCall.logs.length, 1)
   t.is(webhookAfterCall.logs[0].status, 'success')
   t.is(webhookAfterCall.logs[0].metadata.eventObjectId, assetType.id)
@@ -326,6 +262,7 @@ test('creates a webhook with a custom event type', async (t) => {
     permissions: [
       'webhook:create:all',
       'webhook:read:all',
+      'webhookLog:list:all',
       'event:create:all'
     ]
   })
@@ -361,6 +298,13 @@ test('creates a webhook with a custom event type', async (t) => {
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: webhookLogsAfterCall } } = await request(t.context.serverUrl)
+    .get(`/webhook-logs?webhookId=${webhook.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(webhookAfterCall.logs), getIds(webhookLogsAfterCall))
+
   t.is(webhookAfterCall.logs.length, 1)
   t.is(webhookAfterCall.logs[0].status, 'success')
 })
@@ -371,6 +315,7 @@ test('creates a webhook with targetUrl server returning errors', async (t) => {
     permissions: [
       'webhook:create:all',
       'webhook:read:all',
+      'webhookLog:list:all',
       'category:create:all'
     ]
   })
@@ -399,6 +344,13 @@ test('creates a webhook with targetUrl server returning errors', async (t) => {
     .get(`/webhooks/${webhook.id}?logs=`)
     .set(authorizationHeaders)
     .expect(200)
+
+  const { body: { results: webhookLogsAfterCall } } = await request(t.context.serverUrl)
+    .get(`/webhook-logs?webhookId=${webhook.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(webhookAfterCall.logs), getIds(webhookLogsAfterCall))
 
   t.is(webhookAfterCall.logs.length, 1)
   t.is(webhookAfterCall.logs[0].status, 'error')
@@ -468,86 +420,6 @@ test('removes a webhook', async (t) => {
     .get(`/webhooks/${webhook.id}`)
     .set(authorizationHeaders)
     .expect(404)
-})
-
-// need serial to ensure there is no insertion/deletion during pagination scenario
-test.serial('list webhook logs', async (t) => {
-  const authorizationHeaders = await getAccessTokenHeaders({ t, permissions: ['webhookLog:list:all'] })
-
-  await checkCursorPaginationScenario({
-    t,
-    endpointUrl: '/webhook-logs',
-    authorizationHeaders,
-  })
-})
-
-test('list webhook logs with id filter', async (t) => {
-  const authorizationHeaders = await getAccessTokenHeaders({ t, permissions: ['webhookLog:list:all'] })
-
-  const { body: { results: webhookLogs } } = await request(t.context.serverUrl)
-    .get('/webhook-logs')
-    .set(authorizationHeaders)
-    .expect(200)
-
-  const webhookLog = webhookLogs[0]
-
-  const { body: obj } = await request(t.context.serverUrl)
-    .get(`/webhook-logs?id=${webhookLog.id}`)
-    .set(authorizationHeaders)
-    .expect(200)
-
-  checkCursorPaginatedListObject(t, obj)
-  t.is(obj.results.length, 1)
-})
-
-test('list webhook logs with advanced filters', async (t) => {
-  const authorizationHeaders = await getAccessTokenHeaders({ t, permissions: ['webhookLog:list:all'] })
-
-  const now = new Date().toISOString()
-  const minDate = computeDate(now, '-10d')
-
-  const {
-    entry__created: entryWorkflow,
-    message__created: messageWorkflow,
-  } = createdWebhooks
-
-  const params = `createdDate[gte]=${encodeURIComponent(minDate)}` +
-    `&webhookId[]=${entryWorkflow.id}` +
-    `&webhookId[]=${messageWorkflow.id}`
-
-  const { body: obj } = await request(t.context.serverUrl)
-    .get(`/webhook-logs?${params}`)
-    .set(authorizationHeaders)
-    .expect(200)
-
-  const checkResultsFn = (t, webhookLog) => {
-    t.true(webhookLog.createdDate >= minDate)
-    t.true([entryWorkflow.id, messageWorkflow.id].includes(webhookLog.webhookId))
-  }
-
-  checkCursorPaginatedListObject(t, obj, { checkResultsFn })
-})
-
-test('finds a webhook log', async (t) => {
-  const authorizationHeaders = await getAccessTokenHeaders({
-    t,
-    permissions: [
-      'webhookLog:list:all',
-      'webhookLog:read:all'
-    ]
-  })
-
-  const { body: { results: webhookLogs } } = await request(t.context.serverUrl)
-    .get('/webhook-logs')
-    .set(authorizationHeaders)
-    .expect(200)
-
-  const { body: webhookLog } = await request(t.context.serverUrl)
-    .get(`/webhook-logs/${webhookLogs[0].id}`)
-    .set(authorizationHeaders)
-    .expect(200)
-
-  t.is(webhookLog.id, webhookLogs[0].id)
 })
 
 // ////////// //
