+++ test/integration/api/workflow.spec.js
@@ -18,13 +18,14 @@ const { getAccessTokenHeaders } = require('../../auth')
 const {
   getObjectEvent,
   testEventMetadata,
-  computeDate,
 
   checkCursorPaginationScenario,
   checkCursorPaginatedListObject,
 } = require('../../util')
 
 const { encodeBase64 } = require('../../../src/util/encoding')
+
+const getIds = (elements) => elements.map(e => e.id)
 
 let userWebhookUrl
 /* eslint-disable no-template-curly-in-string */
@@ -36,125 +37,6 @@ const isActionLog = log => log.type === 'action'
 const isSkippedLog = log => log.type === 'skipped'
 const isStoppedLog = log => log.type === 'stopped'
 const isNotificationLog = log => log.type === 'notification'
-
-let createdWorkflows
-
-async function createWorkflowLogs (t) {
-  if (createdWorkflows) return createdWorkflows
-
-  const authorizationHeaders = await getAccessTokenHeaders({
-    t,
-    permissions: [
-      'workflowLog:list:all',
-      'workflow:create:all',
-      'category:create:all',
-      'entry:create:all',
-      'message:create:all',
-    ]
-  })
-
-  // should create workflows that listen to events that are not triggered by any tests below
-
-  const { body: categoryWorkflow } = await request(t.context.serverUrl)
-    .post('/workflows')
-    .set(authorizationHeaders)
-    .send({
-      name: 'Workflow for category creation',
-      event: 'category__created',
-      run: { // single-step object allowed
-        endpointMethod: 'PATCH',
-        endpointUri: '/categories/${object.id}',
-        endpointPayload: JSON.stringify({ // simulate real API call (string JSON only)
-          metadata: {
-            updated: true
-          }
-        })
-      }
-    })
-    .expect(200)
-
-  const { body: messageWorkflow } = await request(t.context.serverUrl)
-    .post('/workflows')
-    .set(authorizationHeaders)
-    .send({
-      name: 'Workflow for message creation',
-      event: 'message__created',
-      run: { // single-step object allowed
-        endpointMethod: 'PATCH',
-        endpointUri: '/messages/${object.id}',
-        endpointPayload: JSON.stringify({ // simulate real API call (string JSON only)
-          metadata: {
-            updated: true
-          }
-        })
-      }
-    })
-    .expect(200)
-
-  const { body: entryWorkflow } = await request(t.context.serverUrl)
-    .post('/workflows')
-    .set(authorizationHeaders)
-    .send({
-      name: 'Workflow for entry creation',
-      event: 'entry__created',
-      run: { // single-step object allowed
-        endpointMethod: 'PATCH',
-        endpointUri: '/entries/${object.id}',
-        endpointPayload: JSON.stringify({ // simulate real API call (string JSON only)
-          metadata: {
-            updated: true
-          }
-        })
-      }
-    })
-    .expect(200)
-
-  createdWorkflows = _.keyBy([
-    categoryWorkflow,
-    messageWorkflow,
-    entryWorkflow,
-  ], 'event')
-
-  await request(t.context.serverUrl)
-    .post('/categories')
-    .set(authorizationHeaders)
-    .send({
-      name: 'Some category',
-    })
-    .expect(200)
-
-  await request(t.context.serverUrl)
-    .post('/messages')
-    .set(authorizationHeaders)
-    .send({
-      topicId: 'ast_2l7fQps1I3a1gJYz2I3a',
-      receiverId: 'user-external-id',
-      content: 'Good',
-    })
-    .expect(200)
-
-  await request(t.context.serverUrl)
-    .post('/entries')
-    .set(authorizationHeaders)
-    .send({
-      collection: 'someCollection',
-      locale: 'en-US',
-      name: 'nameExample',
-      fields: {
-        title: 'Random title',
-        content: 'Random content',
-        nestedContent: {
-          random1: {
-            random2: 'hello'
-          },
-          random3: 'bye'
-        }
-      }
-    })
-    .expect(200)
-
-  await new Promise(resolve => setTimeout(resolve, defaultTestDelay))
-}
 
 test.before(async (t) => {
   await before({ name: 'workflow' })(t)
@@ -197,8 +79,6 @@ test.before(async (t) => {
       resolve()
     })
   })
-
-  await createWorkflowLogs(t)
 })
 // test.beforeEach(beforeEach()) // tests are run concurrently
 test.after(async (t) => {
@@ -333,7 +213,8 @@ test('creates several single-step Stelace workflows', async (t) => {
     t,
     permissions: [
       'workflow:create:all',
-      'workflow:read:all'
+      'workflow:read:all',
+      'workflowLog:list:all',
     ]
   })
   const workflowName = 'Single-step Workflow'
@@ -436,6 +317,13 @@ test('creates several single-step Stelace workflows', async (t) => {
     .get(`/workflows/${workflow1.id}?logs=`)
     .set(authorizationHeaders)
     .expect(200)
+
+  const { body: { results: workflowLogsAfterRun1 } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow1.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflow1AfterRun1.logs), getIds(workflowLogsAfterRun1))
 
   const workflow1AfterRun1ErrorLogs = workflow1AfterRun1.logs.filter(isErrorLog)
   const workflow1AfterRun1Actions = workflow1AfterRun1.logs.filter(isActionLog)
@@ -576,6 +464,13 @@ test('creates several single-step Stelace workflows', async (t) => {
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsAfterRun2 } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow1.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflow1AfterRun2.logs), getIds(workflowLogsAfterRun2))
+
   const workflow1AfterRun2ErrorLogs = workflow1AfterRun2.logs.filter(isErrorLog)
   const workflowAfterRun2Notifications = workflow1AfterRun2.logs.filter(isNotificationLog)
   const workflow1AfterRun2Actions = workflow1AfterRun2.logs.filter(isActionLog)
@@ -672,7 +567,8 @@ test('creates multi-step Stelace workflow with API version', async (t) => {
     t,
     permissions: [
       'workflow:create:all',
-      'workflow:read:all'
+      'workflow:read:all',
+      'workflowLog:list:all',
     ]
   })
   const apiVersion = '2019-05-20'
@@ -810,6 +706,13 @@ test('creates multi-step Stelace workflow with API version', async (t) => {
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsAfterDummy } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowAfterDummy.logs), getIds(workflowLogsAfterDummy))
+
   const afterDummyLogs = workflowAfterDummy.logs.filter(
     log => log.metadata.eventObjectId === dummyAssetId
   )
@@ -890,6 +793,13 @@ test('creates multi-step Stelace workflow with API version', async (t) => {
     .get(`/workflows/${workflow.id}?logs=`)
     .set(authorizationHeaders)
     .expect(200)
+
+  const { body: { results: workflowLogsAfterRun } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowAfterRun.logs), getIds(workflowLogsAfterRun))
 
   const workflowAfterRunErrorLogs = workflowAfterRun.logs.filter(isErrorLog)
   const workflowAfterRunActions = workflowAfterRun.logs.filter(isActionLog)
@@ -983,6 +893,7 @@ test('creates multi-step workflow triggered by custom events and calling externa
     permissions: [
       'workflow:create:all',
       'workflow:read:all',
+      'workflowLog:list:all',
       'event:create:all',
       'asset:read:all'
     ]
@@ -1075,6 +986,13 @@ test('creates multi-step workflow triggered by custom events and calling externa
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsAfterRun } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflowCustomEvent.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowCustomEventAfterRun.logs), getIds(workflowLogsAfterRun))
+
   const workflowCustomEventAfterRunErrorLogs = workflowCustomEventAfterRun.logs.filter(isErrorLog)
   const workflowCustomEventAfterRunActions = workflowCustomEventAfterRun.logs.filter(isActionLog)
 
@@ -1115,6 +1033,7 @@ test('keeps filtered workflow running when handleErrors option is enabled in err
       'workflow:create:all',
       'workflow:edit:all',
       'workflow:read:all',
+      'workflowLog:list:all',
       'event:create:all',
       'asset:read:all'
     ]
@@ -1208,6 +1127,13 @@ test('keeps filtered workflow running when handleErrors option is enabled in err
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsHandlingErrorsAfterRun } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflowHandlingErrors.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowHandlingErrorsAfterRun.logs), getIds(workflowLogsHandlingErrorsAfterRun))
+
   const afterRunLogs = workflowHandlingErrorsAfterRun.logs
   const afterRunErrorLogs = afterRunLogs.filter(isErrorLog)
   const afterRunActions = afterRunLogs.filter(l => isActionLog(l) || isSkippedLog(l))
@@ -1280,6 +1206,13 @@ test('keeps filtered workflow running when handleErrors option is enabled in err
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsNotHandlingErrorsAfterRun } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflowHandlingErrors.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowNotHandlingErrorsAfterRun.logs), getIds(workflowLogsNotHandlingErrorsAfterRun))
+
   const notHandlingErrLogs = workflowNotHandlingErrorsAfterRun.logs.filter(
     l => l.runId !== lastAction.runId
   )
@@ -1305,6 +1238,7 @@ test('creates workflow and uses related objects', async (t) => {
     permissions: [
       'workflow:create:all',
       'workflow:read:all',
+      'workflowLog:list:all',
       'transaction:read:all',
       'transaction:edit:all',
       'transaction:config:all'
@@ -1360,6 +1294,13 @@ test('creates workflow and uses related objects', async (t) => {
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsRelatedObjectsAfterRun } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflowRelatedObjects.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowRelatedObjectsAfterRun.logs), getIds(workflowLogsRelatedObjectsAfterRun))
+
   const workflowRelatedObjectsAfterRunErrorLogs = workflowRelatedObjectsAfterRun.logs.filter(isErrorLog)
   const workflowRelatedObjectsAfterRunActions = workflowRelatedObjectsAfterRun.logs.filter(isActionLog)
 
@@ -1394,7 +1335,8 @@ test('accepts nested arrays of literals as endpoint payload parameters', async (
     t,
     permissions: [
       'workflow:read:all',
-      'workflow:create:all'
+      'workflow:create:all',
+      'workflowLog:list:all',
     ]
   })
   const workflowName = 'Workflow test array'
@@ -1462,6 +1404,13 @@ test('accepts nested arrays of literals as endpoint payload parameters', async (
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsTestArrayAfterRun } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflowTestArray.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowTestArrayAfterRun.logs), getIds(workflowLogsTestArrayAfterRun))
+
   const workflowTestArrayAfterRunErrorLogs = workflowTestArrayAfterRun.logs.filter(isErrorLog)
   const workflowTestArrayAfterRunActions = workflowTestArrayAfterRun.logs.filter(isActionLog)
 
@@ -1492,7 +1441,8 @@ test('accepts nested object as endpoint payload parameters', async (t) => {
     t,
     permissions: [
       'workflow:read:all',
-      'workflow:create:all'
+      'workflow:create:all',
+      'workflowLog:list:all',
     ]
   })
   const workflowName = 'Workflow test nested object'
@@ -1553,6 +1503,13 @@ test('accepts nested object as endpoint payload parameters', async (t) => {
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsTestNestedObjectAfterRun } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflowTestNestedObject.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowTestNestedObjectAfterRun.logs), getIds(workflowLogsTestNestedObjectAfterRun))
+
   const workflowTestNestedObjectAfterRunErrorLogs = workflowTestNestedObjectAfterRun.logs.filter(isErrorLog)
   const workflowTestNestedObjectAfterRunActions = workflowTestNestedObjectAfterRun.logs.filter(isActionLog)
 
@@ -1589,6 +1546,7 @@ test('handles filters and logs errors properly when executing Stelace Workflow',
       'workflow:create:all',
       'workflow:edit:all',
       'workflow:read:all',
+      'workflowLog:list:all',
       'user:read:all',
       'user:edit:all'
     ],
@@ -1662,6 +1620,13 @@ test('handles filters and logs errors properly when executing Stelace Workflow',
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsAfterStepsErrors } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowAfterStepsErrors.logs), getIds(workflowLogsAfterStepsErrors))
+
   // Exclude any log due to workflow events from concurrent tests
   const afterStepsErrorsLogs = workflowAfterStepsErrors.logs.filter(
     log => log.metadata.eventObjectId === userId
@@ -1726,6 +1691,13 @@ test('handles filters and logs errors properly when executing Stelace Workflow',
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsWithHandleErrorsEnabled } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowWithHandleErrorsEnabled.logs), getIds(workflowLogsWithHandleErrorsEnabled))
+
   // Step 2 is started but not executed since it has its own error
   t.is(workflowWithHandleErrorsEnabled.stats.nbActionsCompleted, 1)
 
@@ -1780,6 +1752,13 @@ test('handles filters and logs errors properly when executing Stelace Workflow',
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsAfterStep2Fix } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowAfterStep2Fix.logs), getIds(workflowLogsAfterStep2Fix))
+
   const afterStep2FixLogs = workflowAfterStep2Fix.logs.filter(
     log => log.metadata.eventObjectId === userId
   )
@@ -1832,6 +1811,13 @@ test('handles filters and logs errors properly when executing Stelace Workflow',
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsAfterStepsFixed } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowAfterStepsFixed.logs), getIds(workflowLogsAfterStepsFixed))
+
   const workflowAfterStepsFixedErrorLogs = workflowAfterStepsFixed.logs.filter(isErrorLog)
   const workflowAfterStepsFixedActions = workflowAfterStepsFixed.logs.filter(isActionLog)
   const AfterStepsFixedRunId1 = workflowAfterStepsFixedActions[0].runId
@@ -1850,7 +1836,8 @@ test('passes basic security checks', async (t) => {
     permissions: [
       'workflow:create:all',
       'workflow:edit:all',
-      'workflow:read:all'
+      'workflow:read:all',
+      'workflowLog:list:all',
     ]
   })
   const workflowName = 'Evil Workflow'
@@ -1919,6 +1906,13 @@ test('passes basic security checks', async (t) => {
     .set(authorizationHeaders)
     .expect(200)
 
+  const { body: { results: workflowLogsAfterRun } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowAfterRun.logs), getIds(workflowLogsAfterRun))
+
   const workflowAfterRunErrorLogs = workflowAfterRun.logs.filter(isErrorLog)
   const workflowAfterRunActions = workflowAfterRun.logs.filter(isActionLog)
   const workflowAfterRunNotifications = workflowAfterRun.logs.filter(isNotificationLog)
@@ -1977,6 +1971,13 @@ test('passes basic security checks', async (t) => {
     .get(`/workflows/${workflow.id}?logs=`)
     .set(authorizationHeaders)
     .expect(200)
+
+  const { body: { results: workflowLogsWithProcess } } = await request(t.context.serverUrl)
+    .get(`/workflow-logs?workflowId=${workflow.id}`)
+    .set(authorizationHeaders)
+    .expect(200)
+
+  t.deepEqual(getIds(workflowWithProcess.logs), getIds(workflowLogsWithProcess))
 
   const workflowWithProcessErrorLogs = workflowWithProcess.logs.filter(isErrorLog)
   workflowLastError = workflowWithProcessErrorLogs[0]
@@ -2040,86 +2041,6 @@ test('removes a workflow', async (t) => {
     .get(`/workflows/${workflow.id}`)
     .set(authorizationHeaders)
     .expect(404)
-})
-
-// need serial to ensure there is no insertion/deletion during pagination scenario
-test.serial('list workflow logs', async (t) => {
-  const authorizationHeaders = await getAccessTokenHeaders({ t, permissions: ['workflowLog:list:all'] })
-
-  await checkCursorPaginationScenario({
-    t,
-    endpointUrl: '/workflow-logs',
-    authorizationHeaders,
-  })
-})
-
-test('list workflow logs with id filter', async (t) => {
-  const authorizationHeaders = await getAccessTokenHeaders({ t, permissions: ['workflowLog:list:all'] })
-
-  const { body: { results: workflowLogs } } = await request(t.context.serverUrl)
-    .get('/workflow-logs')
-    .set(authorizationHeaders)
-    .expect(200)
-
-  const workflowLog = workflowLogs[0]
-
-  const { body: obj } = await request(t.context.serverUrl)
-    .get(`/workflow-logs?id=${workflowLog.id}`)
-    .set(authorizationHeaders)
-    .expect(200)
-
-  checkCursorPaginatedListObject(t, obj)
-  t.is(obj.results.length, 1)
-})
-
-test('list workflow logs with advanced filters', async (t) => {
-  const authorizationHeaders = await getAccessTokenHeaders({ t, permissions: ['workflowLog:list:all'] })
-
-  const now = new Date().toISOString()
-  const minDate = computeDate(now, '-10d')
-
-  const {
-    category__created: categoryWorkflow,
-    message__created: messageWorkflow,
-  } = createdWorkflows
-
-  const params = `createdDate[gte]=${encodeURIComponent(minDate)}` +
-    `&workflowId[]=${categoryWorkflow.id}` +
-    `&workflowId[]=${messageWorkflow.id}`
-
-  const { body: obj } = await request(t.context.serverUrl)
-    .get(`/workflow-logs?${params}`)
-    .set(authorizationHeaders)
-    .expect(200)
-
-  const checkResultsFn = (t, workflowLog) => {
-    t.true(workflowLog.createdDate >= minDate)
-    t.true([categoryWorkflow.id, messageWorkflow.id].includes(workflowLog.workflowId))
-  }
-
-  checkCursorPaginatedListObject(t, obj, { checkResultsFn })
-})
-
-test('finds a workflow log', async (t) => {
-  const authorizationHeaders = await getAccessTokenHeaders({
-    t,
-    permissions: [
-      'workflowLog:list:all',
-      'workflowLog:read:all'
-    ]
-  })
-
-  const { body: { results: workflowLogs } } = await request(t.context.serverUrl)
-    .get('/workflow-logs')
-    .set(authorizationHeaders)
-    .expect(200)
-
-  const { body: workflowLog } = await request(t.context.serverUrl)
-    .get(`/workflow-logs/${workflowLogs[0].id}`)
-    .set(authorizationHeaders)
-    .expect(200)
-
-  t.is(workflowLog.id, workflowLogs[0].id)
 })
 
 // ////////// //
